errors: fixed `errors_queue.len` type
Make `errors_queue` member `len` of type `size_t`, instead of the wrong
type `unsigned int`.

+++ src/errors.c
@@ -41,8 +41,8 @@
 
 
 static struct errors_queue_tag {
-    unsigned int len;
-    char*        messages[RHD_ERRORS_QUEUE_MAX];
+    size_t len;
+    char*  messages[RHD_ERRORS_QUEUE_MAX];
 } errors_queue = RHD_ERRORS_QUEUE_INIT;
